create models file
