+++ .gitignore
@@ -9,3 +9,4 @@ src/llms/models
 build
 deno.lock
 out
+types
